+++ database/migrations/2025_04_15_090140_create_monev_keluaran_submissions.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('indikator_keluaran_id')->constrained('monev_indikator_keluarans');
             $table->foreignId('user_id')->constrained('users');
             $table->year('tahun');
-            $table->decimal('capaian', 10, 2);
+            $table->integer('capaian');
             $table->tinyInteger('status')->default(0); // 0=pending, 1=approved, 2=rejected
             $table->timestamps();
